+++ src/main/java/it/unipd/mtss/business/bill/Bill.Java
@@ -3,7 +3,7 @@
 // [RAFFAELE] [OLIVIERO] [1216748]
 ////////////////////////////////////////////////////////////////////
 
-package it.unipd.mtss.business;
+package it.unipd.mtss.business.bill;
 import java.util.List;
 import it.unipd.mtss.business.exception.BillException;
 import it.unipd.mtss.model.EItem;
